fix security

+++ src/Controller/CommentController.php
@@ -10,18 +10,11 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 #[Route('/comment')]
 class CommentController extends AbstractController
 {
-    #[Route('/', name: 'comment_index', methods: ['GET'])]
-    public function index(CommentRepository $commentRepository): Response
-    {
-        return $this->render('comment/index.html.twig', [
-            'comments' => $commentRepository->findAll(),
-        ]);
-    }
-
     #[Route('/new', name: 'comment_new', methods: ['GET', 'POST'])]
     public function new(Request $request, PostRepository $postRepository): Response
     {
@@ -46,62 +39,50 @@ class CommentController extends AbstractController
         }
     }
 
-    #[Route('/{id}', name: 'comment_show', methods: ['GET'])]
-    public function show(Comment $comment): Response
-    {
-        return $this->render('comment/show.html.twig', [
-            'comment' => $comment,
-        ]);
-    }
-
-    #[Route('/{id}/edit', name: 'comment_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Comment $comment): Response
-    {
-        $form = $this->createForm(CommentType::class, $comment);
-        $form->handleRequest($request);
-
-        if ($form->isSubmitted() && $form->isValid()) {
-            $comment->setUpdatedAt(new \DateTime());
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('comment_index');
-        }
-
-        return $this->render('comment/edit.html.twig', [
-            'comment' => $comment,
-            'form' => $form->createView(),
-        ]);
-    }
-
+    /**
+     * @IsGranted("ROLE_VALID")
+     */
     #[Route('/{id}', name: 'comment_delete', methods: ['DELETE'])]
-    public function delete(Request $request, CommentRepository $commentRepository, $id): Response
+    public function delete(Request $request, CommentRepository $commentRepository, PostRepository $postRepository, $id): Response
     {
         $comment = $commentRepository->find($id);
+        $userPosts = $postRepository->findBy(['user' => $this->getUser()->getId()]);
 
         if($comment) {
-            if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
-                $entityManager = $this->getDoctrine()->getManager();
-                $entityManager->remove($comment);
-                $entityManager->flush();
+            if(in_array($comment->getPost(), $userPosts) || $this->isGranted('ROLE_ADMIN')) {
+                if ($this->isCsrfTokenValid('delete'.$comment->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($comment);
+                    $entityManager->flush();
+                }
             }
         }
 
         return $this->redirect($request->headers->get('referer'));
     }
 
+    /**
+     * @IsGranted("ROLE_VALID")
+     */
     #[Route('/{id}/validation', name: 'comment_validation', methods: ['POST'])]
-    public function validation(Comment $comment)
+    public function validation(Comment $comment, CommentRepository $commentRepository, PostRepository $postRepository, $id)
     {
+        $comment = $commentRepository->find($id);
         $isValid = $comment->getIsValid() == 1;
+        $userPosts = $postRepository->findBy(['user' => $this->getUser()->getId()]);
 
-        if(!$isValid) {
-            $comment->setIsValid(1);
-        }else {
-            $comment->setIsValid(0);
+        if($comment) {
+            if(in_array($comment->getPost(), $userPosts) || $this->isGranted('ROLE_ADMIN')) {
+
+                if(!$isValid) {
+                    $comment->setIsValid(1);
+                }else {
+                    $comment->setIsValid(0);
+                }
+
+                $this->getDoctrine()->getManager()->flush();
+            }
         }
-
-        $this->getDoctrine()->getManager()->flush();
-
         return $this->redirectToRoute('admin_post_edit', ['id' => $comment->getPost()->getId()]);
     }
