+++ src/Controller/PostController.php
@@ -13,6 +13,7 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\IsGranted;
 
 #[Route('/post')]
 class PostController extends AbstractController
@@ -21,11 +22,11 @@ class PostController extends AbstractController
     public function index(PostRepository $postRepository): Response
     {
         return $this->render('post/index.html.twig', [
-            'posts' => $postRepository->findAll(),
+            'posts' => $postRepository->findBy([], ['id' => 'DESC']),
         ]);
     }
 
-    #[Route('/new', name: 'post_new', methods: ['GET', 'POST'])]
+    #[Route('/new', name: 'post_new', methods: ['POST'])]
     public function new(Request $request): Response
     {
         $post = new Post();
@@ -41,13 +42,8 @@ class PostController extends AbstractController
             $entityManager->persist($post);
             $entityManager->flush();
 
-            return $this->redirectToRoute('post_index');
+            return $this->redirectToRoute('admin_index');
         }
-
-        return $this->render('post/new.html.twig', [
-            'post' => $post,
-            'form' => $form->createView(),
-        ]);
     }
 
     #[Route('/{id}', name: 'post_show', methods: ['GET'])]
@@ -62,27 +58,43 @@ class PostController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}/edit', name: 'post_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Post $post): Response
+    /**
+     * @IsGranted("ROLE_VALID")
+     */
+    #[Route('/{id}/edit', name: 'post_edit', methods: ['POST'])]
+    public function edit(Request $request, Post $post, PostRepository $postRepository, $id): Response
     {
-        $form = $this->createForm(PostType::class, $post);
-        $form->handleRequest($request);
+        $userPosts = $postRepository->findBy(["user" => $this->getUser()->getId()]);
+        $repo = $postRepository->find($id);
 
-        if ($form->isSubmitted() && $form->isValid()) {
-            $post->setUpdatedAt(new \DateTime());
-            $this->getDoctrine()->getManager()->flush();
-
-            return $this->redirectToRoute('admin_index');
+        if($repo) {
+            if(in_array($post, $userPosts) || $this->isGranted('ROLE_ADMIN')) {
+                $form = $this->createForm(PostType::class, $post);
+                $form->handleRequest($request);
+        
+                if ($form->isSubmitted() && $form->isValid()) {
+                    $post->setUpdatedAt(new \DateTime());
+                    $this->getDoctrine()->getManager()->flush();
+                }
+            }
         }
+
+        return $this->redirectToRoute('admin_index');
     }
 
     #[Route('/{id}', name: 'post_delete', methods: ['DELETE'])]
-    public function delete(Request $request, Post $post): Response
+    public function delete(Request $request, Post $post, PostRepository $postRepository, $id): Response
     {
-        if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
-            $entityManager = $this->getDoctrine()->getManager();
-            $entityManager->remove($post);
-            $entityManager->flush();
+        $userPosts = $postRepository->findBy(['user' => $this->getUser()->getId()]);
+
+        if($postRepository->find($id)) {
+            if(in_array($post, $userPosts) || $this->isGranted('ROLE_ADMIN')) {
+                if ($this->isCsrfTokenValid('delete'.$post->getId(), $request->request->get('_token'))) {
+                    $entityManager = $this->getDoctrine()->getManager();
+                    $entityManager->remove($post);
+                    $entityManager->flush();
+                }
+            }
         }
 
         return $this->redirectToRoute('admin_index');
